+++ crates/vm-core/src/virtualization/vm.rs
@@ -17,7 +17,7 @@ pub enum SetUserMemoryRegionFlags {
 pub trait HypervisorVm: Send + Sync {
     fn create_vcpu(
         &self,
-        vcpu_id: usize,
+        vcpu_id: u64,
         mm: Arc<MemoryAddressSpace>,
         vm_exit_handler: Arc<dyn VmExit>,
     ) -> Result<Box<dyn HypervisorVcpu>, VmError>;
